requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
playwright==1.40.0
httpx==0.25.2
requests==2.31.0
pandas==2.1.4
pdfplumber==0.10.3
pytesseract==0.3.10
python-magic==0.4.27
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1
openai==1.3.9